requirements.txt
tk==0.1.0
tkcalendar==1.6.1
babel==2.17.0
pillow==11.2.1
PyMySQL==1.1.1
python-dotenv==1.1.0